Move authorizer_factory out to pkg/auth/configuration
Prepare for remote authentication and authorization to avoid future
curcular dependency.
