+++ pkg/auth/BUILD.bazel
@@ -6,7 +6,6 @@ go_library(
         "any_authorizer.go",
         "authentication_metadata.go",
         "authorizer.go",
-        "authorizer_factory.go",
         "jmespath_expression_authorizer.go",
         "static_authorizer.go",
     ],
@@ -16,7 +15,6 @@ go_library(
         "//pkg/digest",
         "//pkg/otel",
         "//pkg/proto/auth",
-        "//pkg/proto/configuration/auth",
         "//pkg/util",
         "@com_github_jmespath_go_jmespath//:go-jmespath",
         "@io_opentelemetry_go_otel//attribute",
