+++ pkg/auth/configuration/authorizer_factory.go
@@ -1,6 +1,7 @@
-package auth
+package configuration
 
 import (
+	"github.com/buildbarn/bb-storage/pkg/auth"
 	"github.com/buildbarn/bb-storage/pkg/digest"
 	pb "github.com/buildbarn/bb-storage/pkg/proto/configuration/auth"
 	"github.com/buildbarn/bb-storage/pkg/util"
@@ -16,7 +17,7 @@ import (
 type AuthorizerFactory interface {
 	// NewAuthorizerFromConfiguration constructs an authorizer based on
 	// options specified in a configuration message.
-	NewAuthorizerFromConfiguration(configuration *pb.AuthorizerConfiguration) (Authorizer, error)
+	NewAuthorizerFromConfiguration(configuration *pb.AuthorizerConfiguration) (auth.Authorizer, error)
 }
 
 // DefaultAuthorizerFactory constructs deduplicated authorizers based on
@@ -29,15 +30,15 @@ type BaseAuthorizerFactory struct{}
 
 // NewAuthorizerFromConfiguration constructs an authorizer based on
 // options specified in a configuration message.
-func (f BaseAuthorizerFactory) NewAuthorizerFromConfiguration(config *pb.AuthorizerConfiguration) (Authorizer, error) {
+func (f BaseAuthorizerFactory) NewAuthorizerFromConfiguration(config *pb.AuthorizerConfiguration) (auth.Authorizer, error) {
 	if config == nil {
 		return nil, status.Error(codes.InvalidArgument, "Authorizer configuration not specified")
 	}
 	switch policy := config.Policy.(type) {
 	case *pb.AuthorizerConfiguration_Allow:
-		return NewStaticAuthorizer(func(in digest.InstanceName) bool { return true }), nil
+		return auth.NewStaticAuthorizer(func(in digest.InstanceName) bool { return true }), nil
 	case *pb.AuthorizerConfiguration_Deny:
-		return NewStaticAuthorizer(func(in digest.InstanceName) bool { return false }), nil
+		return auth.NewStaticAuthorizer(func(in digest.InstanceName) bool { return false }), nil
 	case *pb.AuthorizerConfiguration_InstanceNamePrefix:
 		trie := digest.NewInstanceNameTrie()
 		for _, i := range policy.InstanceNamePrefix.AllowedInstanceNamePrefixes {
@@ -47,13 +48,13 @@ func (f BaseAuthorizerFactory) NewAuthorizerFromConfiguration(config *pb.Authori
 			}
 			trie.Set(instanceNamePrefix, 0)
 		}
-		return NewStaticAuthorizer(trie.ContainsPrefix), nil
+		return auth.NewStaticAuthorizer(trie.ContainsPrefix), nil
 	case *pb.AuthorizerConfiguration_JmespathExpression:
 		expression, err := jmespath.Compile(policy.JmespathExpression)
 		if err != nil {
 			return nil, util.StatusWrapWithCode(err, codes.InvalidArgument, "Failed to compile JMESPath expression")
 		}
-		return NewJMESPathExpressionAuthorizer(expression), nil
+		return auth.NewJMESPathExpressionAuthorizer(expression), nil
 	default:
 		return nil, status.Error(codes.InvalidArgument, "Unknown authorizer configuration")
 	}
@@ -62,7 +63,7 @@ func (f BaseAuthorizerFactory) NewAuthorizerFromConfiguration(config *pb.Authori
 type deduplicatingAuthorizerFactory struct {
 	base AuthorizerFactory
 	// Keys are protojson-encoded pb.AuthorizerConfigurations
-	known map[string]Authorizer
+	known map[string]auth.Authorizer
 }
 
 // NewDeduplicatingAuthorizerFactory creates a new AuthorizerFactory
@@ -71,12 +72,12 @@ type deduplicatingAuthorizerFactory struct {
 func NewDeduplicatingAuthorizerFactory(base AuthorizerFactory) AuthorizerFactory {
 	return &deduplicatingAuthorizerFactory{
 		base:  base,
-		known: make(map[string]Authorizer),
+		known: make(map[string]auth.Authorizer),
 	}
 }
 
 // NewAuthorizerFromConfiguration creates an Authorizer based on the passed configuration.
-func (af *deduplicatingAuthorizerFactory) NewAuthorizerFromConfiguration(config *pb.AuthorizerConfiguration) (Authorizer, error) {
+func (af *deduplicatingAuthorizerFactory) NewAuthorizerFromConfiguration(config *pb.AuthorizerConfiguration) (auth.Authorizer, error) {
 	keyBytes, err := protojson.Marshal(config)
 	key := string(keyBytes)
 	if err != nil {
